app/src/main/java/com/ashutosh1234ojha/unittestapp/data/local/ShoppingDao.kt
package com.ashutosh1234ojha.unittestapp.data.local

import androidx.lifecycle.LiveData
import androidx.room.*

/**
 * Created by Ashutosh Ojha on 24,October,2021
 */
@Dao
interface ShoppingDao {
    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertShoppingItem(shoppingItem: ShoppingItem)

    @Delete
    suspend fun deleteShoppingItem(shoppingItem: ShoppingItem)

    @Query("SELECT * FROM shopping_items")
    fun observeAllShoppingItems(): LiveData<List<ShoppingItem>>

    @Query("SELECT SUM(price * amount) FROM shopping_items")
    fun observeTotalPrice(): LiveData<Float>
}